+++ frontend2/src/index.tsx
@@ -1,8 +1,23 @@
 /* @refresh reload */
 import { render } from 'solid-js/web'
-import './index.css'
-import App from './App.tsx'
+import { Router, Route } from "@solidjs/router";
+import '~/index.css'
+// import App from './App.tsx'
+
+import Home from "~/routes/home.tsx";
+import Dashboard from "~/routes/dashboard.tsx";
 
 const root = document.getElementById('root')
 
-render(() => <App />, root!)
+if (!root) {
+  throw new Error("Wrapper div not found");
+}
+// render(() => <App />, root!)
+render(
+  () => (
+    <Router>
+      <Route path="/" component={Home} />
+      <Route path="/dashboard" component={Dashboard} />
+    </Router>
+  ), root
+);